requirements.txt
openai
langchain
streamlit==1.29.0
python-dotenv
pyPDF2


-e .